requirement.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.0
pyparsing==2.4.7
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
python-dateutil==2.8.1
six==1.15.0
